+++ tests/suites/connection/ConnectionTest.php
@@ -10,8 +10,12 @@ declare(strict_types=1);
 namespace WebSocket\Test\Connection;
 
 use PHPUnit\Framework\TestCase;
+use Phrity\Net\Context;
 use Phrity\Net\Mock\SocketStream;
-use Phrity\Net\Mock\Stack\ExpectSocketStreamTrait;
+use Phrity\Net\Mock\Stack\{
+    ExpectContextTrait,
+    ExpectSocketStreamTrait,
+};
 use Psr\Log\NullLogger;
 use Stringable;
 use WebSocket\Connection;
@@ -37,6 +41,7 @@ use WebSocket\Middleware\Callback;
  */
 class ConnectionTest extends TestCase
 {
+    use ExpectContextTrait;
     use ExpectSocketStreamTrait;
 
     public function setUp(): void
@@ -56,6 +61,7 @@ class ConnectionTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();
@@ -105,6 +111,9 @@ class ConnectionTest extends TestCase
 
         $this->assertFalse($connection->isConnected());
 
+        $this->expectSocketStreamGetContext();
+        $this->assertInstanceOf(Context::class, $connection->getContext());
+
         unset($connection);
     }
 
@@ -114,6 +123,7 @@ class ConnectionTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();
@@ -135,6 +145,7 @@ class ConnectionTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();
@@ -176,6 +187,7 @@ class ConnectionTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();
@@ -210,6 +222,7 @@ class ConnectionTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();
@@ -234,6 +247,7 @@ class ConnectionTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();
@@ -258,6 +272,7 @@ class ConnectionTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();
@@ -282,6 +297,7 @@ class ConnectionTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();