+++ tests/suites/middleware/ProcessStackTest.php
@@ -11,7 +11,10 @@ namespace WebSocket\Test\Middleware;
 
 use PHPUnit\Framework\TestCase;
 use Phrity\Net\Mock\SocketStream;
-use Phrity\Net\Mock\Stack\ExpectSocketStreamTrait;
+use Phrity\Net\Mock\Stack\{
+    ExpectContextTrait,
+    ExpectSocketStreamTrait,
+};
 use Psr\Log\NullLogger;
 use WebSocket\Connection;
 use WebSocket\Message\Text;
@@ -26,6 +29,7 @@ use WebSocket\Middleware\{
  */
 class ProcessStackTest extends TestCase
 {
+    use ExpectContextTrait;
     use ExpectSocketStreamTrait;
 
     public function setUp(): void
@@ -45,6 +49,7 @@ class ProcessStackTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();
@@ -94,6 +99,7 @@ class ProcessStackTest extends TestCase
 
         $this->expectSocketStream();
         $this->expectSocketStreamGetMetadata();
+        $this->expectContext();
         $stream = new SocketStream($temp);
 
         $this->expectSocketStreamGetLocalName();